backend/requirements.txt
pandas
mysql-connector-python
python-dotenv
langchain
langgraph
langchain-community
langchain-experimental
langchain-core
langchain-openai